Don't treat ipv4 as ipv6 with 16-byte representation
The Go IP lib seems to represent ipv4 internally in the 16-byte
ipv4 mapped address space. So the To16 method will simply convert
a 4-byte ipv4 to this representation. It seems like ipsource is
using To16 to test whether an IP is a v6 address, but this will also
include all v4 IPs ... it's not clear whether this is intentional.